Make Spring-test.js (mostly) work again
Updating the test to reflect the latest API change of only animating
shallow objects, and obligatorily passing the spring config.

Disabling the tests for unmounting for now; we're not allowing Motion
(previously, Spring)'s `style` (previously, `endValue`) to be a
function anymore, so we need another way to test this rather than
destroying the component through rendering a new one to the root inside
the endValue callback.

@bsansouci