Handle `undefined` and `null` in style objects
I am upgrading from 0.2.7 to 0.3.0 and I ran into an error:

> Uncaught TypeError: Cannot read property 'config' of undefined

This happens when I am using a `TransitionMotion` with a style object
that has a key who's value is `undefined`, e.g.:

  {
    opacity: spring(0),
    extraText: undefined,
  }

To make react-motion easier to work with, I have added guards in a
number of places.

Fixes #180.