Types.js: turn key type from any to string only
See previous and next few commits.

We're doing a small rewrite of the `styles` merging algorithm of TM
that'll cache `TransitionStyle`/`TransitionPlainStyle` `key` as object
keys internally. This means the user provided `key` must be a string
since js objects keys cast stuff to that.

(Also adding a TODO following #273.)